Add liquid comments and fix display of month year
Slight typo on the previous commit in formatting the date display
for the nested list contained. Only month was showing instead of
month year because I'd quoted both values.

Also, added comments in for the liquid template code to explain
what is going on.

Finally: Can't believe this works! Damn clever!