Styling for Archive filter
Adds in styling for the archive page art/code filter (not great, like the
filter, but it will do). Also changes to suit the change to the #archive
id be moved from the ul, up to the article element itself. And removing
the border from the article element on the archive page.