Can't truncate you idiot as causes mark-up errors
I wondered why my Archive page wouldn't display correctly when I added a
load of posts, but then realised that truncating a post would lead to lots
of open tags/elements and, especially with posts with lists, this caused
a right mess with the lists used on the archive page.

Also, there is really no need to truncate, it's 'truncated' visually
anyway.