Fix video link
Due to export from Tumblr (could be due to how post was originally made on
Tumblr or changes on Tumblr itself over time) rather than just a link to
video being supplied the embedded code was picked up. And due to the way my
templates are setup this just means it trys to embded it again.