Change markup to site standard (to pick up styles)
Also, change shared post quantity to 10, as per Index and Category pages.
Add paragraph with link to full RSS feed for shared items

Note, can't expose this in the UI yet (will add a link inbetween Archive
and Elsewhere) because need to set "Access-Control-Allow-Origin" header
on sub-domain that Fever is hosted on. Damn CORs.