Same trick, use truncate and capture instead of contains
Same issue with other templates. The verison of Liquid on Github
does not seem to have the "contains" function I used to check the
video source to see if it was Vimeo or Youtube. So use "truncate"
instead and test against this. Requires capturing it into a
variable though as can't nest this truncate within the if
statement.

However, with all this said and done, I don't think the iframes
are working within the feed anyway.""