Remove category from permalink
I never really realised how crappy that was especially when posts
have multiple categories. This way should make it easier to use
relative links to images, etc.

EDIT: Actually strike that. I can't fathom how to decently use
relative links in Jekyll. There is the option of using
{{ site.baseurl }} as a prefix to urls, but I can't use this in
the YAML section anyway. So might as well just stick to absolute urls
to images, etc (Especially when considering URLs in the atom feed)

But still like the idea of removing categories from the URLs.