Use baseurl rather than hard coding absolute url
Had a brainwave and figured out how to have relative urls in
photo posts and yet still get everything hunkydory: I.e. don't
try to use liquid tags in the yaml section, rather modify the
photo include so that it only prepends the baseurl to images that
don't start with http.